restaurant_bot/requirements.txt
pyTelegramBotAPI==4.7.0
python-dotenv==1.0.0
